requirements.txt
openai==1.57.0
ffmpeg-python==0.2.0
jiwer==3.0.4
pandas==2.2.3
matplotlib==3.9.3
seaborn==0.13.2
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.67.1
yt-dlp==2024.12.6
requests==2.32.3
beautifulsoup4==4.12.3